Conflicts:
	Repository/src/main/java/sportstats/repository/GameRepository.java
	Rest/src/main/java/sportstats/rest/GameResource.java
	Service/src/main/java/sportstats/service/GameService.java

+++ Domain/src/test/java/tests/domain/GameTest.java
@@ -88,7 +88,7 @@ public class GameTest {
     @Test
     public void spectatorTest() {
         Game game = new Game();
-        Long spectators = 100L;
+        int spectators = 100;
 
         game.setSpectators(spectators);
         assertEquals(spectators, game.getSpectators());
